requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
Faker==26.0.0
django-filter==24.2
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
